Fixed showing fps
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@5132 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/utils/orcaimageview/component.cpp
@@ -65,7 +65,7 @@ Component::start()
 
     // connect signal from timer to slot of widget to update
     QObject::connect(imageQueue_, SIGNAL(imagePushed()), viewer, SLOT(updateGL()));
-    QObject::connect(viewer, SIGNAL(fpsChanged(int)), fpsDisplay, SLOT(setNum(int)));
+    QObject::connect(viewer, SIGNAL(fpsChanged(double)), fpsDisplay, SLOT(setNum(double)));
 
     // start network thread
     mainThread_ = new MainThread(imageQueue_, context() );
